aider: Modified `run.sh` script to mount the directories so that the files are changed on the host, and updated `solution.py` to write the output CSV to the output directory.
# Aider chat conversation:

USER: run-script needs to mount the directories so the files are actually changed on the host

ASSISTANT: To modify the `run.sh` script to mount the directories so that the files are changed on the host, we need to add volume mounts to the `docker run` command. This will ensure that the input files and the output file are accessible from the host file system.

Here is the *SEARCH/REPLACE block* for the `run.sh` script:

run.sh
```bash
<<<<<<< SEARCH
# Run the Docker container
docker run --rm --name solution-container solution-image
=======
# Run the Docker container with volume mounts
docker run --rm --name solution-container -v "$(pwd)/input:/usr/src/app/input" -v "$(pwd)/output:/usr/src/app/output" solution-image
>>>>>>> REPLACE
```

This change will mount the `input` directory from the current working directory on the host to `/usr/src/app/input` inside the container, and the `output` directory from the current working directory on the host to `/usr/src/app/output` inside the container. Make sure to create the `output` directory on the host if it does not exist.

Additionally, we need to update the `solution.py` file to write the output CSV to the `output` directory.

solution.py
```python
<<<<<<< SEARCH
# Write the merged DataFrame to a new CSV file
merged_df.to_csv('./out.csv', index=False)
=======
# Write the merged DataFrame to a new CSV file in the output directory
merged_df.to_csv('./output/out.csv', index=False)
>>>>>>> REPLACE
```

Make sure to create the `output` directory on your host system before running the `run.sh` script, or the script will fail because Docker will not be able to mount a non-existent directory.